LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/popovers/popover-top-layer-nesting.tentative-expected.txt

Single popover=auto ancestor
Single popover=manual ancestor
Nested popover=auto ancestors
Nested popover=auto ancestors, target is outer
Top layer inside of nested element


Harness Error (TIMEOUT), message = null

FAIL Single popover=auto ancestor with dialog assert_equals: Incorrect behavior expected true but got false
PASS Single popover=auto ancestor with dialog, top layer element *is* a popover
FAIL Single popover=auto ancestor with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Single popover=auto ancestor with fullscreen, top layer element *is* a popover
FAIL Single popover=manual ancestor with dialog assert_equals: Incorrect behavior expected true but got false
PASS Single popover=manual ancestor with dialog, top layer element *is* a popover
FAIL Single popover=manual ancestor with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Single popover=manual ancestor with fullscreen, top layer element *is* a popover
FAIL Nested popover=auto ancestors with dialog assert_equals: Incorrect behavior expected true but got false
PASS Nested popover=auto ancestors with dialog, top layer element *is* a popover
FAIL Nested popover=auto ancestors with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Nested popover=auto ancestors with fullscreen, top layer element *is* a popover
FAIL Nested popover=auto ancestors, target is outer with dialog assert_equals: Incorrect behavior expected true but got false
PASS Nested popover=auto ancestors, target is outer with dialog, top layer element *is* a popover
FAIL Nested popover=auto ancestors, target is outer with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Nested popover=auto ancestors, target is outer with fullscreen, top layer element *is* a popover
FAIL Top layer inside of nested element with dialog assert_equals: Incorrect behavior expected true but got false
PASS Top layer inside of nested element with dialog, top layer element *is* a popover
TIMEOUT Top layer inside of nested element with fullscreen Test timed out
NOTRUN Top layer inside of nested element with fullscreen, top layer element *is* a popover
